requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
PyJWT==2.8.0
